requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.4
APScheduler==3.10.4
fastapi==0.103.1
uvicorn[standard]==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.3.0